evdev: don't truncate event time to 32 bits
This causes a bunch of "your system is too slow" messages in e.g. the various
gesture tests.

Fixes 95a72990
Fixes #601

(cherry picked from commit 04dc67e09250cdf275b139ac9ab0b9a118a7de29)

+++ src/evdev.c
@@ -1024,7 +1024,7 @@ evdev_note_time_delay(struct evdev_device *device,
 {
 	struct libinput *libinput = evdev_libinput_context(device);
 	uint32_t tdelta;
-	uint32_t eventtime = input_event_time(ev);
+	uint64_t eventtime = input_event_time(ev);
 
 	/* if we have a current libinput_dispatch() snapshot, compare our
 	 * event time with the one from the snapshot. If we have more than
